+++ applications/engine/pom.xml
@@ -10,9 +10,9 @@
     <relativePath>../../pom.xml</relativePath>
   </parent>
 
-  <artifactId>parity-system</artifactId>
+  <artifactId>parity-engine</artifactId>
 
-  <name>Parity Trading System</name>
+  <name>Parity Matching Engine</name>
 
   <dependencies>
     <dependency>
@@ -64,10 +64,10 @@
         <groupId>org.apache.maven.plugins</groupId>
         <artifactId>maven-shade-plugin</artifactId>
         <configuration>
-          <outputFile>parity-system.jar</outputFile>
+          <outputFile>parity-engine.jar</outputFile>
           <transformers>
             <transformer implementation="org.apache.maven.plugins.shade.resource.ManifestResourceTransformer">
-              <mainClass>com.paritytrading.parity.system.TradingSystem</mainClass>
+              <mainClass>com.paritytrading.parity.engine.MatchingEngine</mainClass>
             </transformer>
           </transformers>
         </configuration>
