+++ applications/engine/src/main/java/com/paritytrading/parity/engine/Order.java
@@ -1,4 +1,4 @@
-package com.paritytrading.parity.system;
+package com.paritytrading.parity.engine;
 
 import com.paritytrading.parity.match.OrderBook;
