+++ libraries/util/src/main/java/com/paritytrading/parity/util/package-info.java
@@ -1,4 +1,4 @@
 /**
- * Parity Utilities contain support functions used by Parity Trading System.
+ * Parity Utilities contain support functions used by the trading system.
  */
 package com.paritytrading.parity.util;
